image display & refactor
